% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoParams.R
\name{alien_synthesis}
\alias{alien_synthesis}
\title{function generating new alien species}
\usage{
alien_synthesis(trait_range, n = 1)
}
\arguments{
\item{trait_range}{a dataframe containing the
parameters to change and their accepted values.
Default is NULL, leading to the use of the NS_params
object}

\item{n}{An integer to know the number of aliens to
generate}
}
\description{
Function uses a range of trait to produce alien
Default range of trait is based on North Sea
parameters but can be supplied by user
}
