% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoParams.R
\name{evoParams}
\alias{evoParams}
\title{Set up parameters for an evolutionary model}
\usage{
evoParams(
  no_sp = 11,
  min_w_inf = 10,
  max_w_inf = 10^4,
  min_w = 10^(-3),
  max_w = max_w_inf,
  eta = 10^(-0.6),
  min_w_mat = min_w_inf * eta,
  no_w = log10(max_w_inf/min_w) * 20 + 1,
  min_w_pp = 1e-10,
  w_pp_cutoff = min_w_mat,
  n = 2/3,
  p = n,
  lambda = 2.05,
  r_pp = 0.1,
  kappa = 0.005,
  alpha = 0.4,
  h = 40,
  beta = 100,
  sigma = 1.3,
  f0 = 0.6,
  fc = 0.25,
  ks = NA,
  gamma = NA,
  zeta = 0.2,
  ext_mort_prop = 0,
  R_factor = 4,
  gear_names = "knife_edge_gear",
  knife_edge_size = 1000,
  RDD = "extinctionRDD",
  egg_size_scaling = FALSE,
  resource_scaling = FALSE,
  perfect_scaling = FALSE,
  specificParams = NULL,
  interactionMatrix = NULL,
  ea_int = 0,
  ca_int = 0
)
}
\arguments{
\item{no_sp}{The number of species in the model.}

\item{min_w_inf}{The asymptotic size of the smallest species in the
community. This will be rounded to lie on a grid point.}

\item{max_w_inf}{The asymptotic size of the largest species in the community.
This will be rounded to lie on a grid point.}

\item{min_w}{The size of the the egg of the smallest species. This also
defines the start of the community size spectrum.}

\item{max_w}{The largest size in the model. By default this is set to the
largest asymptotic size \code{max_w_inf}. Setting it to something larger
only makes sense if you plan to add larger species to the model later.}

\item{eta}{Ratio between maturity size and asymptotic size of a species.
Ignored if \code{min_w_mat} is supplied. Default is 10^(-0.6),
approximately 1/4.}

\item{min_w_mat}{The maturity size of the smallest species. Default value is
\code{eta * min_w_inf}. This will be rounded to lie on a grid point.}

\item{no_w}{The number of size bins in the community spectrum. These bins
will be equally spaced on a logarithmic scale. Default value is such that
there are 50 bins for each factor of 10 in weight.}

\item{min_w_pp}{The smallest size of the resource spectrum. By default this
is set to the smallest value at which any of the consumers can feed.}

\item{w_pp_cutoff}{The largest size of the resource spectrum. Default value
is max_w_inf unless \code{perfect_scaling = TRUE} when it is Inf.}

\item{n}{Scaling exponent of the maximum intake rate.}

\item{p}{Scaling exponent of the standard metabolic rate. By default this is
equal to the exponent \code{n}.}

\item{lambda}{Exponent of the abundance power law.}

\item{r_pp}{Growth rate parameter for the resource spectrum.}

\item{kappa}{Coefficient in abundance power law.}

\item{alpha}{The assimilation efficiency of the community.}

\item{h}{Maximum food intake rate.}

\item{beta}{Preferred predator prey mass ratio.}

\item{sigma}{Width of prey size preference.}

\item{f0}{Expected average feeding level. Used to set \code{gamma}, the
coefficient in the search rate. Ignored if \code{gamma} is given
explicitly.}

\item{fc}{Critical feeding level. Used to determine \code{ks} if it is not given
explicitly.}

\item{ks}{Standard metabolism coefficient. If not provided, default will be
calculated from critical feeding level argument \code{fc}.}

\item{gamma}{Volumetric search rate. If not provided, default is determined
by \code{\link[=get_gamma_default]{get_gamma_default()}} using the value of \code{f0}.}

\item{zeta}{...}

\item{ext_mort_prop}{The proportion of the total mortality that comes from
external mortality, i.e., from sources not explicitly modelled. A number in
the interval [0, 1).}

\item{R_factor}{The factor such that \code{R_max = R_factor * R}, where \code{R_max}
is the maximum reproduction rate allowed and \code{R} is the steady-state
reproduction rate. Thus the larger \code{R_factor} the less the impact of the
density-dependence.}

\item{gear_names}{The names of the fishing gears for each species. A
character vector, the same length as the number of species.}

\item{knife_edge_size}{The minimum size at which the gear or gears select
fish. A single value for each gear or a vector with one value for each
gear.}

\item{RDD}{...}

\item{egg_size_scaling}{If TRUE, the egg size is a constant fraction of the
maximum size of each species. This fraction is \code{min_w / min_w_inf}. If
FALSE, all species have the egg size \code{w_min}.}

\item{resource_scaling}{If TRUE, the carrying capacity for larger resource
is reduced to compensate for the fact that fish eggs and larvae are
present in the same size range.}

\item{perfect_scaling}{If TRUE then parameters are set so that the community
abundance, growth before reproduction and death are perfect power laws. In
particular all other scaling corrections are turned on.}

\item{specificParams}{...}

\item{interactionMatrix}{...}

\item{ea_int}{...}

\item{ca_int}{...}
}
\value{
An object of type \code{MizerParams}
}
\description{
this will either set up new params as usual or convert Mizer params to MizerEvo params if given a mizer.params object (TODO)
}
\examples{
\dontrun{
params <- newTraitParams()
sim <- project(params, t_max = 5, effort = 0)
plotSpectra(sim)
}
}
\concept{functions for setting up models}
