% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoParams.R
\name{evoProject}
\alias{evoProject}
\title{Project}
\usage{
evoProject(
  initCondition = NULL,
  params = NULL,
  t_max = 100,
  dt = 0.1,
  mutation = 2,
  trait = "w_mat",
  initPool = 0,
  initSpread = 5,
  alien = 0,
  saveFolder = file.path(tempdir(), "simTemp"),
  effort = 0
)
}
\arguments{
\item{params}{...}

\item{t_max}{...}

\item{mutation}{...}

\item{saveFolder}{...}
}
\value{
...
}
\description{
Project
}
\examples{
\donttest{
saveFolder <- file.path(tempdir(), "simTemp")
dir.create(saveFolder)
params <- evoParams(no_sp = 5, RDD = "extinctionRDD")
sim <- evoProject(params = params, t_max = 300, mutation = 5,
                  saveFolder = saveFolder)
plot(sim)
}
}
