% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoParams.R
\name{extinctionRDD}
\alias{extinctionRDD}
\title{Stock recruitment relationship enabling extinction of species}
\usage{
extinctionRDD(rdi, species_params, ...)
}
\arguments{
\item{rdi}{...}

\item{species_params}{...}

\item{...}{...}
}
\description{
when a species reaches an abundance threshold its recruitment gets disabled, species is later removed
For now each species acts as its own
TODO at the moment I am using rdi to apply the threhold but it should be done on n directly
}
