% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{mizerEvo}
\alias{mizerEvo}
\alias{mizerEvo-package}
\title{mizerEvo: Simulating Evolutionary Processes with Mizer}
\description{
This is an extension package for the mizer package
(https://sizespectrum.org/mizer/) that contains additional features
to ...
}
