% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotevoFeeding}
\alias{plotevoFeeding}
\title{Plot feeding level}
\usage{
plotevoFeeding(
  object,
  time_range = max(as.numeric(dimnames(object@n)$time)),
  species = T,
  throughTime = F,
  start = 1000,
  every = 1000,
  print_it = T,
  returnData = F,
  save_it = F,
  nameSave = "Feeding.png"
)
}
\value{
A plot
}
\description{

}
\examples{
\donttest{
}
}
\seealso{
\link{plotting_functions},

Other plotting functions: 
\code{\link{plotDynamics}()},
\code{\link{plotPredMort}()},
\code{\link{plotPredRate}()},
\code{\link{plotSS}()},
\code{\link{plotSpawn}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoMortality}()},
\code{\link{plotevoTrait}()}
}
\concept{plotting functions}
