% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotevoMortality}
\alias{plotevoMortality}
\title{Plot total mortality}
\usage{
plotevoMortality(
  object,
  time_range = max(as.numeric(dimnames(object@n)$time)),
  print_it = TRUE,
  returnData = F,
  comments = T
)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim}}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step.}

\item{print_it}{A boolean value that displays comments. Default is TRUE}

\item{returnData}{A boolean value that controls the output. Either a dataframe
ready for ggplot (TRUE) or the plot itself (FALSE). Default is FALSE}
}
\value{
A plot
}
\description{

}
\examples{
\donttest{
}
}
\seealso{
\link{plotting_functions},

Other plotting functions: 
\code{\link{plotDynamics}()},
\code{\link{plotPredMort}()},
\code{\link{plotPredRate}()},
\code{\link{plotSS}()},
\code{\link{plotSpawn}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoFeeding}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoTrait}()}
}
\concept{plotting functions}
