% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotevoTrait}
\alias{plotevoTrait}
\title{Plot trait evolution through time}
\usage{
plotevoTrait(
  object,
  SpIdx = NULL,
  Normalisation = F,
  returnData = T,
  traitID = NULL
)
}
\value{
A plot
}
\description{

}
\examples{
\donttest{
}
}
\seealso{
\link{plotting_functions},

Other plotting functions: 
\code{\link{plotDynamics}()},
\code{\link{plotPredMort}()},
\code{\link{plotPredRate}()},
\code{\link{plotSS}()},
\code{\link{plotSpawn}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoFeeding}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoMortality}()}
}
\concept{plotting functions}
