% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPredMort}
\alias{plotPredMort}
\title{Plot predation mortality rate of each species against size}
\usage{
plotPredMort(object, species = NULL, time_range, highlight = NULL)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim}}

\item{species}{Plot only a specific species and its phenotypes
Default is NULL}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step.}

\item{highlight}{...}
}
\value{
A plot
}
\description{
After running a projection, plot the predation mortality rate of each species
by size. The mortality rate is averaged over the specified time range (a
single value for the time range can be used to plot a single time step).
}
\examples{
\donttest{
params <- suppressMessages(newMultispeciesParams(NS_species_params_gears, inter))
sim <- project(params, effort=1, t_max=20, t_save = 2, progress_bar = FALSE)
plotPredMort(sim)
plotPredMort(sim, time_range = 10:20)
}
}
\seealso{
\link{plotting_functions},  \code{\link[=getPredMort]{getPredMort()}}

Other plotting functions: 
\code{\link{plotDynamics}()},
\code{\link{plotPredRate}()},
\code{\link{plotSS}()},
\code{\link{plotSpawn}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoFeeding}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoMortality}()},
\code{\link{plotevoTrait}()}
}
\concept{plotting functions}
