% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSpawn}
\alias{plotSpawn}
\title{Plot spawn}
\usage{
plotSpawn(
  object,
  time_range = max(as.numeric(dimnames(object@n)$time)),
  species = T,
  print_it = T,
  returnData = F,
  save_it = F,
  nameSave = "Spawn.png",
  ...
)
}
\value{
A plot
}
\description{

}
\examples{
\donttest{
}
}
\seealso{
\link{plotting_functions},

Other plotting functions: 
\code{\link{plotDynamics}()},
\code{\link{plotPredMort}()},
\code{\link{plotPredRate}()},
\code{\link{plotSS}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoFeeding}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoMortality}()},
\code{\link{plotevoTrait}()}
}
\concept{plotting functions}
