% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSS}
\alias{plotSS}
\title{Plot size spectrumms of species and phenotypes}
\usage{
plotSS(
  object,
  time_range = max(as.numeric(dimnames(object@n)$time)),
  min_w = min(object@params@w)/100,
  ylim = c(NA, NA),
  biomass = TRUE,
  print_it = TRUE,
  species = TRUE,
  community = FALSE,
  save_it = FALSE,
  nameSave = "SizeSpectrum.png",
  returnData = FALSE
)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim}}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step.}

\item{min_w}{...}

\item{ylim}{...}

\item{biomass}{...}

\item{print_it}{A boolean value that displays comments. Default is TRUE}

\item{species}{Plot only a specific species and its phenotypes
Default is NULL}

\item{community}{...}

\item{save_it}{...}

\item{nameSave}{...}

\item{returnData}{A boolean value that controls the output. Either a dataframe
ready for ggplot (TRUE) or the plot itself (FALSE). Default is FALSE}
}
\value{
A plot
}
\description{

}
\examples{
\donttest{
}
}
\seealso{
\link{plotting_functions},

Other plotting functions: 
\code{\link{plotDynamics}()},
\code{\link{plotPredMort}()},
\code{\link{plotPredRate}()},
\code{\link{plotSpawn}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoFeeding}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoMortality}()},
\code{\link{plotevoTrait}()}
}
\concept{plotting functions}
