% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoParams.R
\name{tempFun}
\alias{tempFun}
\title{tempFun is a function that takes temperature parameters (temperature, t_ref, t_d) and physiological parameters (w, Ea, c_a)
and returns a scalar based on the Padfield (2016) temperature performance equation. This will be upgraded in the future with
a deactivation portion with Anna Gardmark's equation once it's published}
\usage{
tempFun(w, temperature, t_ref, Ea, c_a)
}
\description{
tempFun is a function that takes temperature parameters (temperature, t_ref, t_d) and physiological parameters (w, Ea, c_a)
and returns a scalar based on the Padfield (2016) temperature performance equation. This will be upgraded in the future with
a deactivation portion with Anna Gardmark's equation once it's published
}
