% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDynamics}
\alias{plotDynamics}
\title{Plot biomass variations through time of species and phenotypes}
\usage{
plotDynamics(
  object,
  time_range = c(min(as.numeric(dimnames(object@n)$time)),
    max(as.numeric(dimnames(object@n)$time))),
  phenotype = TRUE,
  species = NULL,
  trait = NULL,
  SpIdx = NULL,
  print_it = T,
  returnData = F,
  save_it = F,
  nameSave = "Biomass.png",
  ylimit = c(NA, NA)
)
}
\value{
A plot
}
\description{

}
\examples{
\donttest{
}
}
\seealso{
\link{plotting_functions},

Other plotting functions: 
\code{\link{plotPredMort}()},
\code{\link{plotPredRate}()},
\code{\link{plotSS}()},
\code{\link{plotSpawn}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoFeeding}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoMortality}()},
\code{\link{plotevoTrait}()}
}
\concept{plotting functions}
