% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotevoMortality}
\alias{plotevoMortality}
\title{Plot total mortality}
\usage{
plotevoMortality(
  object,
  time_range = max(as.numeric(dimnames(object@n)$time)),
  print_it = TRUE,
  returnData = F,
  comments = T
)
}
\value{
A plot
}
\description{

}
\examples{
\donttest{
}
}
\seealso{
\link{plotting_functions},

Other plotting functions: 
\code{\link{plotDynamics}()},
\code{\link{plotPredMort}()},
\code{\link{plotPredRate}()},
\code{\link{plotSS}()},
\code{\link{plotSpawn}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoFeeding}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoTrait}()}
}
\concept{plotting functions}
