% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPredRate}
\alias{plotPredRate}
\title{Plot predation rate}
\usage{
plotPredRate(
  object,
  time_range = max(as.numeric(dimnames(object@n)$time)),
  species = T,
  ylim = c(NA, NA),
  print_it = T,
  returnData = F,
  save_it = F,
  nameSave = "PredRate.png"
)
}
\arguments{
\item{object}{An object of class \linkS4class{MizerSim}}

\item{time_range}{The time range (either a vector of values, a vector of min
and max time, or a single value) to average the abundances over. Default is
the final time step.}

\item{species}{Plot only a specific species and its phenotypes
Default is NULL}

\item{ylim}{...}

\item{print_it}{A boolean value that displays comments. Default is TRUE}

\item{returnData}{A boolean value that controls the output. Either a dataframe
ready for ggplot (TRUE) or the plot itself (FALSE). Default is FALSE}

\item{save_it}{...}

\item{nameSave}{...}
}
\value{
A plot
}
\description{

}
\examples{
\donttest{
}
}
\seealso{
\link{plotting_functions},

Other plotting functions: 
\code{\link{plotDynamics}()},
\code{\link{plotPredMort}()},
\code{\link{plotSS}()},
\code{\link{plotSpawn}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoFeeding}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoMortality}()},
\code{\link{plotevoTrait}()}
}
\concept{plotting functions}
