% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSS}
\alias{plotSS}
\title{Plot size spectrumms of species and phenotypes}
\usage{
plotSS(
  object,
  time_range = max(as.numeric(dimnames(object@n)$time)),
  min_w = min(object@params@w)/100,
  ylim = c(NA, NA),
  biomass = TRUE,
  print_it = TRUE,
  species = TRUE,
  community = FALSE,
  save_it = FALSE,
  nameSave = "SizeSpectrum.png",
  returnData = FALSE,
  ...
)
}
\value{
A plot
}
\description{

}
\examples{
\donttest{
}
}
\seealso{
\link{plotting_functions},

Other plotting functions: 
\code{\link{plotDynamics}()},
\code{\link{plotPredMort}()},
\code{\link{plotPredRate}()},
\code{\link{plotSpawn}()},
\code{\link{plotStarvation}()},
\code{\link{plotevoFeeding}()},
\code{\link{plotevoGrowth}()},
\code{\link{plotevoMortality}()},
\code{\link{plotevoTrait}()}
}
\concept{plotting functions}
